Stable Version. Added deprecated method for copy and error handler for manual init
