+++ app/settings/routes.py
@@ -1,21 +1,19 @@
 import subprocess
-import json
+
 import time
 from os import path, popen
 
 from flask_login import login_required
 from werkzeug.security import safe_join
-from pydbus import SystemBus
-from gi.repository import GLib
 
 from app.forms.network import NetworkForm
 from app.forms.sensors import LoraConfigForm
 
 from flask import render_template, flash, current_app, request, redirect, url_for, jsonify, session
 
-from app.utils import sys_uptime, sys_date, sys_ram, sys_cpu_avg, sys_disk, sys_wired_network_config, \
-    SysConfig, sys_service_restart, db_size, db_clean, sys_auto_timezone, sys_reboot, sys_poweroff, \
-    sys_wireless_network_config, sys_manage_ip_route, sys_wireless_config_clear
+from app.utils import (sys_uptime, sys_date, sys_ram, sys_cpu_avg, sys_disk, sys_wired_network_config, \
+                       SysConfig, sys_service_restart, db_size, db_clean, sys_auto_timezone, sys_reboot, sys_poweroff, \
+                       ModemControl)
 
 
 def routes(bp):
@@ -242,7 +240,9 @@ def routes(bp):
                 conn_config_input = ('internet', 4, '', '')
             modem = ModemControl()
 
-            modem.modem_add_connection(conn_config_input)
+            connection_status = modem.modem_add_connection(conn_config_input)
+            session['current_bearer'] = connection_status
+
         return redirect(url_for('settings.modem_settings'))
 
     @bp.route("/switch", methods=['POST'])
@@ -255,7 +255,8 @@ def routes(bp):
         if state:
             modem.modem_delete_connection()
         else:
-            modem.modem_add_connection(('internet', 4, '', ''))
+            connection_status = modem.modem_add_connection(('internet', 4, '', ''))
+            session['current_bearer'] = connection_status
         return redirect(url_for('settings.modem_settings'))
 
     @bp.route("/ussd_request", methods=['POST'])
@@ -283,8 +284,24 @@ def routes(bp):
         dns_status = network.net_config_read('DNS')
 
         ussd_status = modem.modem_ussd_status()
+        current_bearer = session.get('current_bearer', str())
+
+        if '/org/freedesktop/ModemManager1/Bearer' in current_bearer:
+            modem_cur_bearer_prop = modem.modem_get_current_bearer_properties(
+                'Properties', current_bearer)
+            modem_current_apn = modem_cur_bearer_prop['apn']
+            if modem_current_apn == '':
+                modem_current_apn = 'internet'
+
+            modem_current_user = modem_cur_bearer_prop['user']
+            modem_current_pass = modem_cur_bearer_prop['password']
+        else:
+            modem_current_apn = '[Default]'
+            modem_current_user = ''
+            modem_current_pass = ''
 
         if request.method == 'POST':
+
             input_request = request.form.get('input_form')
 
             if input_request == 'ussd_cancel':
@@ -299,232 +316,21 @@ def routes(bp):
 
         else:
             response = False
+
+        output_form = {'modem_status': modem_code_status,
+                       'show_modem': show_modem,
+                       'response': response,
+                       'nw_form': nw_form,
+                       'is_modem_settings': True,
+                       'name_status': name_status,
+                       'address_status': ip_status,
+                       'gateway_status': gw_status,
+                       'dns_status': dns_status,
+                       'ussd_status': ussd_status,
+                       'current_bearer': current_bearer,
+                       'modem_current_apn': modem_current_apn,
+                       'modem_current_user': modem_current_user,
+                       'modem_current_password': modem_current_pass}
+
         return render_template("/settings/modem.html",
-                               modem_status=modem_code_status,
-                               show_modem=show_modem,
-                               response=response,
-                               is_modem_settings=True,
-                               nw_form=nw_form,
-                               name_status=name_status,
-                               address_status=ip_status,
-                               gateway_status=gw_status,
-                               dns_status=dns_status,
-                               ussd_status=ussd_status)
-
-    class ModemControl:
-
-        current_bearer = str()
-        current_name = str()
-
-        def modem_check_state(self):
-            try:
-                obj_current_modem = self.modem_current()
-                modem_state = obj_current_modem['org.freedesktop.DBus.Properties'].Get(
-                    'org.freedesktop.ModemManager1.Modem', 'State')
-                return modem_state
-            except:
-                return 'Modem is missing'
-        def modem_add_connection_nm(self):
-            bus = SystemBus()
-            nm = bus.get("org.freedesktop.NetworkManager", "/org/freedesktop/NetworkManager")
-            modem_path = self.modem_current_nm()
-            nm.AddAndActivateConnection({'connection': {'id': GLib.Variant.new_string('modem'),
-                                                        'type': GLib.Variant.new_string('gsm')},
-                                         'gsm': {'apn': GLib.Variant.new_string('internet')}},
-                                        modem_path, '/')
-            # Detect all modem in system
-
-        def modem_add_connection(self, config_input: tuple[str, int, str, str]):
-            bus = SystemBus()
-            obj_current_modem = self.modem_current()
-
-            ports = obj_current_modem['org.freedesktop.DBus.Properties'].Get('org.freedesktop.ModemManager1.Modem',
-                                                                             'Ports')
-            simple_connect = obj_current_modem['org.freedesktop.ModemManager1.Modem.Simple']
-            bearer_path = simple_connect.Connect({'apn': GLib.Variant.new_string(config_input[0]),
-                                                  'ip-type': GLib.Variant.new_int32(int(config_input[1])),
-                                                  'user': GLib.Variant.new_string(config_input[2]),
-                                                  'password': GLib.Variant.new_string(config_input[3])})
-            obj_current_modem['org.freedesktop.ModemManager1.Modem'].Enable('true')
-            self.current_bearer = bus.get("org.freedesktop.ModemManager1", bearer_path)
-            ip_dict = self.current_bearer['org.freedesktop.DBus.Properties'].Get('org.freedesktop.ModemManager1.Bearer',
-                                                                                 'Ip4Config')
-            self.current_name = [port[0] for port in ports if "ww" in port[0]][0]
-            # port[0] возвращает список из одного элемента
-
-            config_input = (self.current_name, ip_dict['address'], ip_dict['gateway'],
-                            (ip_dict['dns1'], ip_dict['dns2']))
-            # TODO Потенциальная ошибка, если DNS будет один экземпляр
-            sys_wireless_network_config(config_input)
-            sys_service_restart('systemd-networkd')
-            e = sys_manage_ip_route(self.current_name, 500)
-            return self.current_name
-
-        def modem_delete_connection(self):
-            bus = SystemBus()
-            obj_current_modem = self.modem_current()
-            ports = obj_current_modem['org.freedesktop.DBus.Properties'].Get('org.freedesktop.ModemManager1.Modem',
-                                                                             'Ports')
-            self.current_name = [port[0] for port in ports if "wwp" in port[0]][0]
-            simple_connect = obj_current_modem['org.freedesktop.ModemManager1.Modem.Simple']
-
-            e = sys_manage_ip_route(self.current_name, 500, "delete")
-            simple_connect.Disconnect('/')
-            sys_wireless_config_clear()
-            sys_service_restart('systemd-networkd')
-
-            # obj_current_modem['org.freedesktop.ModemManager1.Modem'].Enable('false')
-            return self.current_name
-
-        @staticmethod
-        def modem_current():
-            bus = SystemBus()
-            try:
-                # Root object
-                obj_root = bus.get('.ModemManager1', '/org/freedesktop/ModemManager1')
-                # Modem
-                obj_current_modem = bus.get(
-                    '.ModemManager1',
-                    list(
-                        obj_root['org.freedesktop.DBus.ObjectManager'].GetManagedObjects()
-                    )
-                        # last connected modem
-                    [-1]
-                )
-                return obj_current_modem
-            except:
-                return False
-
-        @staticmethod
-        def modem_system_scan():
-            bus = SystemBus()
-            try:
-                # Root object
-                obj_root = bus.get('.ModemManager1', '/org/freedesktop/ModemManager1')
-                obj_root.ScanDevices()
-                return True
-            except:
-                print('Modem not found')
-                return False
-
-        @staticmethod
-        def modem_current_nm():
-            try:
-                bus = SystemBus()
-                nm = bus.get("org.freedesktop.NetworkManager", "/org/freedesktop/NetworkManager")
-                devices = nm.GetDevices()
-                for device_path in devices:
-                    device = bus.get("org.freedesktop.NetworkManager", device_path)
-                    if device.DeviceType == 8:
-                        return device_path
-                return False
-            except:
-                return False
-
-        def modem_ussd_request(self, ussd_code):
-            """Метод, который позволяет посылать USSD запросы. Не может отменять текущую сессию запросов."""
-            try:
-                obj_current_modem = self.modem_current()
-                # USSD session
-                ussd = obj_current_modem['org.freedesktop.ModemManager1.Modem.Modem3gpp.Ussd']
-                ussd_request = str(ussd.Initiate(ussd_code))
-                return ussd_request
-            except:
-                print('Modem not found')
-                return False
-
-        def modem_ussd_status(self):
-            """Метод, возвращающий код состояния сессии USSD. MM_MODEM_3GPP_USSD_SESSION_STATE_UNKNOWN = 0
- - Unknown state;
-MM_MODEM_3GPP_USSD_SESSION_STATE_IDLE = 1
- - No active session;
-MM_MODEM_3GPP_USSD_SESSION_STATE_ACTIVE = 2
- - A session is active and the mobile is waiting for a response;
-MM_MODEM_3GPP_USSD_SESSION_STATE_USER_RESPONSE = 3
- - The network is waiting for the client's response."""
-            try:
-                obj_current_modem = self.modem_current()
-                ussd_status = obj_current_modem['org.freedesktop.DBus.Properties'].Get(
-                    'org.freedesktop.ModemManager1.Modem.Modem3gpp.Ussd', 'State')
-                return ussd_status
-            except:
-                return False
-
-        def modem_ussd_response(self, ussd_code):
-            obj_current_modem = self.modem_current()
-            ussd = obj_current_modem['org.freedesktop.ModemManager1.Modem.Modem3gpp.Ussd']
-            ussd.Respond(ussd_code)
-
-        def modem_ussd_network_request(self):
-            obj_current_modem = self.modem_current()
-            ussd_response = obj_current_modem['org.freedesktop.DBus.Properties'].Get(
-                'org.freedesktop.ModemManager1.Modem.Modem3gpp.Ussd',
-                'NetworkRequest')
-            return ussd_response
-
-        def modem_ussd_cancel(self):
-            try:
-                obj_current_modem = self.modem_current()
-                # USSD session
-                ussd = obj_current_modem['org.freedesktop.ModemManager1.Modem.Modem3gpp.Ussd']
-                ussd.Cancel()
-                return 'OK'
-            except:
-                return 'Unknown Error'
-
-        def modem_apn_set(self, apn_input):
-            """ Метод, который позволяет настроить APN для текущего профиля"""
-            try:
-                obj_current_modem = self.modem_current()
-                apn_set = obj_current_modem['org.freedesktop.ModemManager1.Modem.Modem3gpp.ProfileManager']
-                apn_set.Set({'profile-id': GLib.Variant.new_int32(1), 'apn': GLib.Variant.new_string(str(apn_input))})
-                response = apn_set.List()
-                return response
-
-            except:
-                return 'False'
-
-        def modem_info(self):
-            # DBUS lib tutorial https://github.com/LEW21/pydbus/blob/master/doc/tutorial.rst
-            try:
-                obj_current_modem = self.modem_current()
-                current_modem = obj_current_modem['org.freedesktop.ModemManager1.Modem']
-
-                # Values
-                info = dict()
-
-                if any(slot != '/' for slot in current_modem.SimSlots):
-                    bus = SystemBus()
-
-                    # SIM card
-                    obj_sim = bus.get('.ModemManager1', current_modem.Sim)
-                    current_sim = obj_sim['org.freedesktop.ModemManager1.Sim']
-
-                    info['manufacturer'] = str(current_modem.Manufacturer)
-                    info['model'] = str(current_modem.Model)
-                    info['operator'] = str(current_sim.OperatorName)
-
-                    # Modem3gpp
-                    modem3gpp = obj_current_modem['org.freedesktop.ModemManager1.Modem.Modem3gpp']
-
-                    info['imei'] = str(modem3gpp.Imei)
-
-                    info['active'] = bool(current_sim.Active)
-                    info['signal'] = int(current_modem.SignalQuality[0])
-                    info['simId'] = str(current_sim.SimIdentifier)
-                    info['imsi'] = str(current_sim.Imsi)
-
-                    # New info
-                    info['primary port'] = str(current_modem.PrimaryPort)
-
-                else:
-                    info['sim'] = bool(False)
-
-                return info
-            except:
-                print('Modem not found')
-                return False
-
-        def getter(self):
-            """ Метод для безопасного возврата информации о модеме """
-            return self.modem_info()
+                               form=output_form)
