+++ app/static/js/select_options_button_handler.js
@@ -2,11 +2,12 @@ var secondary_button = document.getElementById("cnl-btn");
 var input_value = document.getElementById("input_value");
 var con_switch = document.getElementById('connectionSwitch_1');
 
-secondary_button.addEventListener('click', () => {
-input_value.value = "ussd_cancel";
-alert("USSD session canceled");
-});
-
+if (secondary_button != null){
+    secondary_button.addEventListener('click', () => {
+    input_value.value = "ussd_cancel";
+    alert("USSD session canceled");
+    });
+}
 
 
 con_switch.addEventListener('click', () => {
@@ -44,12 +45,19 @@ for(let i = 0; i < copy_btn.length; i ++)
     copy_btn[i].addEventListener('click', () => {
     // + String(copy_btn_click.getAttribute('value')));
    //+ String(copy_btn_click.getAttribute('value')));
-    let info = text_field.textContent;
+
     if (window.isSecureContext) {
+        let info = text_field.textContent;
         navigator.clipboard.writeText(`${info}`).then();
     }
     else {
+        var range = document.createRange();
+        range.selectNodeContents(text_field);
+        var selection = window.getSelection();
+        selection.removeAllRanges();
+        selection.addRange(range);
         document.execCommand('copy');
+        selection.removeAllRanges();
     }
     change_svg_btn.setAttribute("href", "#check2");
     setTimeout(() =>{change_svg_btn.setAttribute("href", "#clipboard")}, 5000)
