+++ app/templates/settings/modem.html
@@ -16,7 +16,7 @@
                     <form method="post" action="{{ url_for('settings.modem_settings') }}">
                         <input class="form-check-input" type="checkbox" role="switch"
                                name="connectionSwitch" id="connectionSwitch_1"
-                               {% if modem_status==11 %} checked {% endif %}>
+                               {% if form['modem_status'] == 11 %} checked {% endif %}>
                         <label class="form-check-label" for="connectionSwitch_1">Connection off/on</label>
                     </form>
 
@@ -83,6 +83,10 @@
                     <div class="form-group mt-3">
                         <button class="btn btn-primary" type="submit">Отправить</button>
                     </div>
+
+                    <div class="form-group mt-3">
+                        {{ form['connection_status'] | safe }}
+                    </div>
                 </form>
 
             </div>
@@ -114,13 +118,13 @@
             </form>
             <div class="card-body">
 
-                <div class="row mt-2">
+                <div class="d-flex mt-2">
 
-                    <div class="col-md-6">
-                        <p>Connection Status: {{ ussd_status | safe }}</p>
+                    <div class="col justify-content-start align-items-start">
+                        <p>Connection Status: {{ form['ussd_status'] | safe }}</p>
                     </div>
 
-                    <div class="col-md-6 justify-content-end">
+                    <div class="col d-flex justify-content-md-end">
                         <button class="btn mt-0 me-0 copy-btn" id="copy_btn_2" value="2">
                             <svg class="bi opacity-75 theme-icon" aria-hidden="true">
                                 <use id="copy_btn_svg_2" href="#clipboard"></use>
@@ -130,13 +134,11 @@
 
                 </div>
 
-                <div class="container d-flex justify-content-center">
-                    <pre id="copy_text_field_2">
-<p>{% if response %}{{response | safe }} {% endif %}</p>
+                <div class="row justify-content-center">
+                    <pre type="text"  id="copy_text_field_2">
+{% if form['response'] %}{{form['response'] | safe }} {% endif %}
                     </pre>
                 </div>
-
-
 
             </div>
         </div>
@@ -160,10 +162,10 @@
             <div class="card-header">
                 Текущая информация
             </div>
-            {% if show_modem %}
+            {% if form['show_modem'] %}
             <table class="table table-striped table-borderless text-center" style="margin: auto;">
 
-            {% for Key, Value in show_modem.items() %}
+            {% for Key, Value in form['show_modem'].items() %}
                 <tr>
                     <th scope = "row">{{ Key | safe }}: </th>
                     <td>{{ Value | safe }}</td>
@@ -172,7 +174,7 @@
 
             </table>
             {% else %}
-                <h1 align="center">Модем не найден!</h1>
+                <h1>Модем не найден!</h1>
             {% endif %}
         </div>
     </div>
@@ -198,11 +200,14 @@
 
                 <div class="container d-flex justify-content-center">
                     <pre id="copy_text_field_1">
-<span>State: {{ modem_status | safe }} </span>
-<span>Name: {{ name_status | safe }} </span>
-<span>Address: {{ address_status | safe }} </span>
-<span>Gateway: {{ gateway_status | safe }} </span>
-<span>DNS: {{ dns_status | safe }} </span>
+<span>State: {{ form['modem_status'] | safe }} </span>
+<span>Name: {{ form['name_status'] | safe }} </span>
+<span>APN: {{ form['modem_current_apn'] | safe }} </span>
+<span>User: {{ form['modem_current_user'] | safe }} </span>
+<span>Password: {{ form['modem_current_password'] | safe }} </span>
+<span>Address: {{ form['address_status'] | safe }} </span>
+<span>Gateway: {{ form['gateway_status'] | safe }} </span>
+<span>DNS: {{ form['dns_status'] | safe }} </span>
                     </pre>
                 </div>
